+++ extension/cart/updateProductsInCart_v1.js
@@ -62,6 +62,6 @@ module.exports = async function (context, input) {
       }
     ))
   } catch (err) {
-    return { messages: handleCartError(err) }
+    return { messages: await handleCartError(err, checkoutCartItems, cartId, context) }
   }
 }