+++ extension/helper/cart.js
@@ -53,12 +53,21 @@ function extractVariantId (product) {
  * @param {Object} err.errors
  * @param {Object} err.errors.line_items
  * @param {{code: string, message: string}[]} err.errors.line_items.[errorType]
+ * @param {Array} checkoutCartItems
+ * @param {int} cartId
+ * @param {SDKContext} context
  *
  * @return {{code: string, message: string, type: string}[]}
  * @throws {Error} If err does not have an errors or error.line_items property
  */
-function handleCartError (err) {
+async function handleCartError (err, checkoutCartItems, cartId, context) {
   if (!err || !err.errors || !err.errors.line_items) throw err
+
+  const itemsToDelete = getOutOfStockLineItemIds(err.errors.line_items).sort((a, b) => b - a)
+
+  if (itemsToDelete.length > 0) {
+    await fixCheckoutQuantities(checkoutCartItems, itemsToDelete, err, cartId, context)
+  }
 
   const errorMessages = []
   Object.values(err.errors.line_items).forEach(errorsPerLineItem => {
@@ -83,4 +92,60 @@ function handleCartError (err) {
   return errorMessages
 }
 
-module.exports = { clearCart, updateCart, extractVariantId, handleCartError }
+/**
+ *
+ * @param {Object} lineItems
+ * @returns {Array}
+ */
+function getOutOfStockLineItemIds (lineItems) {
+  const itemsToDelete = []
+
+  for (let itemId in lineItems) {
+    Object.entries(lineItems[itemId]).forEach(([errorType, errors]) => {
+      if (errors.find(error => error.code === 'not_enough_in_stock' &&
+          error.options &&
+          error.options.remaining === 0
+      )) {
+        itemsToDelete.push(parseInt(itemId))
+      }
+    })
+  }
+
+  return itemsToDelete
+}
+
+/**
+ * @param {Array} checkoutCartItems
+ * @param {Array} itemsToDelete
+ * @param {Error} error
+ * @param {Object} error.errors
+ * @param {Object} error.errors.line_items
+ * @param {int} cartId
+ * @param {SDKContext} context
+ */
+function fixCheckoutQuantities (checkoutCartItems, itemsToDelete, error, cartId, context) {
+  const Shopify = require('../lib/shopify.api.js')(context.config, context.log)
+
+  for (let itemId of itemsToDelete) {
+    checkoutCartItems.splice(itemId, 1)
+    delete error.errors.line_items[itemId]
+  }
+
+  const productData = {
+    'checkout': {
+      'line_items': checkoutCartItems
+    }
+  }
+
+  return new Promise((resolve) => {
+    Shopify.put('/admin/checkouts/' + cartId + '.json', productData, function (err) {
+      if (err) {
+        context.log.error(`Could not fix quantities for checkout ${cartId}`)
+      }
+
+      resolve()
+    })
+  })
+}
+
+module.exports = { clearCart, updateCart, extractVariantId, handleCartError, getOutOfStockLineItemIds }